requirements.txt
agno 
autogen
groq
yfinance
python-dotenv
duckduckgo-search